+++ src/components/UserDetail/Loans.test.js
@@ -40,6 +40,7 @@ describe('Loans', () => {
       barcode: 0,
       id: 0,
     },
+    formatMessage: jest.fn(),
   };
   const renderComponent = (props) => {
     const combined = {