+++ src/components/UserDetail/UserDetail.js
@@ -4,6 +4,7 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import {
   FormattedMessage,
+  injectIntl,
 } from 'react-intl';
 
 import {
@@ -57,6 +58,10 @@ class UserDetail extends React.Component {
     id: PropTypes.string.isRequired,
     label: PropTypes.node,
     settings: PropTypes.arrayOf(PropTypes.object).isRequired,
+    intl: PropTypes.shape({
+      formatMessage: PropTypes.func.isRequired,
+    }).isRequired,
+    ariaLabel: PropTypes.node,
     resources: PropTypes.shape({
       patronGroups: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object),
@@ -76,12 +81,20 @@ class UserDetail extends React.Component {
   };
 
   getUserValue = (user) => {
+    const {
+      ariaLabel,
+      intl: {
+        formatMessage,
+      },
+    } = this.props;
+    const userAriaLabel = ariaLabel || formatMessage({ id: 'ui-checkout.borrower.ariaLabel' });
     const path = `/users/view/${user.id}`;
 
     return (
       <span>
         <Link
           className={css.marginRight}
+          aria-label={userAriaLabel}
           to={path}
         >
           <strong data-test-check-out-patron-full-name>
@@ -93,7 +106,16 @@ class UserDetail extends React.Component {
           tagName="strong"
         />
         {' '}
-        {user.barcode ? (<Link to={path}>{user.barcode}</Link>) : <NoValue />}
+        {
+          user.barcode ?
+            <Link
+              aria-label={userAriaLabel}
+              to={path}
+            >
+              {user.barcode}
+            </Link> :
+            <NoValue />
+        }
       </span>
     );
   };
@@ -107,16 +129,18 @@ class UserDetail extends React.Component {
       settings,
       renderLoans,
       stripes,
+      intl: {
+        formatMessage,
+      },
     } = this.props;
-
     const patronGroups = (resources.patronGroups || {}).records || [];
     const patronGroup = patronGroups[0] || {};
     const statusVal = (get(user, ['active'], '') ? 'ui-checkout.active' : 'ui-checkout.inactive');
-
     const profilePictureLink = user?.personal?.profilePictureLink;
     const profilePicturesEnabled = Boolean(settings.length) && settings[0].enabled;
     const hasViewProfilePicturePerm = stripes.hasPerm('ui-users.profile-pictures.view');
     const displayProfilePicture = profilePicturesEnabled && hasViewProfilePicturePerm;
+
     return (
       <div id={id}>
         <div>
@@ -166,6 +190,7 @@ class UserDetail extends React.Component {
             resources={resources}
             stripes={stripes}
             user={user}
+            formatMessage={formatMessage}
           />
         )}
       </div>
@@ -173,4 +198,4 @@ class UserDetail extends React.Component {
   }
 }
 
-export default withStripes(UserDetail);
+export default withStripes(injectIntl(UserDetail));