+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change history for ui-checkout
 
 ## [11.1.0] In progress
-
+* React v19: refactor away from default props for functional components. Refs UICHKOUT-909.
 
 ## [11.0.2] (https://github.com/folio-org/ui-checkout/tree/v11.0.2) (2024-11-30)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v11.0.1...v11.0.2)