+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [11.1.0] In progress
 * React v19: refactor away from default props for functional components. Refs UICHKOUT-909.
+* Fix accessibility issues. Refs UICHKOUT-933.
 
 ## [11.0.2] (https://github.com/folio-org/ui-checkout/tree/v11.0.2) (2024-11-30)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v11.0.1...v11.0.2)